+++ rollup.config.js
@@ -18,11 +18,6 @@ export default [
 			typescript({ tsconfig: './tsconfig.json' }), // so Rollup can convert TypeScript to JavaScript
             terser(), //minify,
             banner2(() => `/* ${pkg.repository.url} v${pkg.version} */\n`),
-            copy({
-                targets: [
-                  { src: pkg.browser, dest: 'docs/examples/assets/' },
-                ]
-            }),
 		]
 	},
 	// browser-friendly UMD build (basic)
@@ -38,6 +33,12 @@ export default [
 			typescript({ tsconfig: './tsconfig.json' }), // so Rollup can convert TypeScript to JavaScript
             terser(), //minify,
             banner2(() => `/* ${pkg.repository.url} v${pkg.version} */\n`),
+            // copy the full version deferred
+            copy({
+                targets: [
+                  { src: pkg.browser, dest: 'docs/examples/assets/' },
+                ]
+            }),
 		]
 	},