+++ src/Quadtree.ts
@@ -278,6 +278,7 @@ export class Quadtree<ObjectsType extends Rectangle|Circle|Line|Indexable> {
 
     /**
      * Remove an object from the tree.
+     * If you have to remove many objects, consider clearing the entire tree and rebuilding it or use the `fast` flag to cleanup after the last removal.
      * @beta
      * 
      * @example 
@@ -289,21 +290,19 @@ export class Quadtree<ObjectsType extends Rectangle|Circle|Line|Indexable> {
      * ```
      * 
      * @example Bulk fast removals and final cleanup:
-     * ```typescript
+     * ```javascript
      * const tree = new Quadtree({ width: 100, height: 100 });
-     * const rects: Rectangle[] = [];
+     * const rects = [];
      *  for(let i=0; i<20; i++) {
      *    rects[i] = new Rectangle({ x: 25, y: 25, width: 50, height: 50 });
      *    tree.insert(rects[i]);
      *  }
      *  for(let i=rects.length-1; i>0; i--) {
-     *    //fast=true – just remove the object, keep empty subnodes
+     *    //fast=true – just remove the object (may leaves vacant subnodes)
      *    //fast=false – cleanup empty subnodes (default)
      *    const fast = (i !== 0);
      *    tree.remove(rects[i], fast); 
      *  }
-     * 
-     * tree.remove(rects[0], false); 
      * ```
      * 
      * @param obj - Object to be removed.
@@ -330,6 +329,49 @@ export class Quadtree<ObjectsType extends Rectangle|Circle|Line|Indexable> {
         }
 
         return (indexOf !== -1);
+    }
+
+    /**
+     * Update an object already in the tree (shorthand for remove and insert).
+     * If you have to update many objects, consider clearing and rebuilding the 
+     * entire tree or use the `fast` flag to cleanup after the last update.
+     * @beta
+     * 
+     * @example 
+     * ```typescript
+     * const tree = new Quadtree({ width: 100, height: 100, maxObjects: 1 });
+     * const rect1 = new Rectangle({ x: 25, y: 25, width: 10, height: 10 });
+     * const rect2 = new Rectangle({ x: 25, y: 25, width: 10, height: 10 });
+     * tree.insert(rect1);
+     * tree.insert(rect2);
+     * rect1.x = 75;
+     * rect1.y = 75;
+     * tree.update(rect1);
+     * ```
+     * @example Bulk fast update and final cleanup:
+     * ```javascript
+     * const tree = new Quadtree({ width: 100, height: 100 });
+     * const rects = [];
+     *  for(let i=0; i<20; i++) {
+     *    rects[i] = new Rectangle({ x: 20, y: 20, width: 20, height: 20 });
+     *    tree.insert(rects[i]);
+     *  }
+     *  for(let i=rects.length-1; i>0; i--) {
+     *    rects[i].x = 20 + Math.random()*60;
+     *    rects[i].y = 20 + Math.random()*60;
+     *    //fast=true – just re-insert the object (may leaves vacant subnodes)
+     *    //fast=false – cleanup empty subnodes (default)
+     *    const fast = (i !== 0);
+     *    tree.update(rects[i], fast); 
+     *  }
+     * ```
+     * 
+     * @param obj - Object to be updated.
+     * @param fast - Set to true to increase performance temporarily by preventing cleanup of empty subnodes (optional, default: false).
+     */
+    update(obj: ObjectsType, fast=false): void {
+        this.remove(obj, fast);
+        this.insert(obj);
     }
 
     /**