+++ types/Quadtree.d.ts
@@ -162,6 +162,7 @@ export declare class Quadtree<ObjectsType extends Rectangle | Circle | Line | In
     retrieve(obj: Rectangle | Circle | Line | Indexable): ObjectsType[];
     /**
      * Remove an object from the tree.
+     * If you have to remove many objects, consider clearing the entire tree and rebuilding it or use the `fast` flag to cleanup after the last removal.
      * @beta
      *
      * @example
@@ -173,21 +174,19 @@ export declare class Quadtree<ObjectsType extends Rectangle | Circle | Line | In
      * ```
      *
      * @example Bulk fast removals and final cleanup:
-     * ```typescript
+     * ```javascript
      * const tree = new Quadtree({ width: 100, height: 100 });
-     * const rects: Rectangle[] = [];
+     * const rects = [];
      *  for(let i=0; i<20; i++) {
      *    rects[i] = new Rectangle({ x: 25, y: 25, width: 50, height: 50 });
      *    tree.insert(rects[i]);
      *  }
      *  for(let i=rects.length-1; i>0; i--) {
-     *    //fast=true – just remove the object, keep empty subnodes
+     *    //fast=true – just remove the object (may leaves vacant subnodes)
      *    //fast=false – cleanup empty subnodes (default)
      *    const fast = (i !== 0);
      *    tree.remove(rects[i], fast);
      *  }
-     *
-     * tree.remove(rects[0], false);
      * ```
      *
      * @param obj - Object to be removed.
@@ -195,6 +194,45 @@ export declare class Quadtree<ObjectsType extends Rectangle | Circle | Line | In
      * @returns Weather or not the object was removed from THIS node (no recursive check).
      */
     remove(obj: ObjectsType, fast?: boolean): boolean;
+    /**
+     * Update an object already in the tree (shorthand for remove and insert).
+     * If you have to update many objects, consider clearing and rebuilding the
+     * entire tree or use the `fast` flag to cleanup after the last update.
+     * @beta
+     *
+     * @example
+     * ```typescript
+     * const tree = new Quadtree({ width: 100, height: 100, maxObjects: 1 });
+     * const rect1 = new Rectangle({ x: 25, y: 25, width: 10, height: 10 });
+     * const rect2 = new Rectangle({ x: 25, y: 25, width: 10, height: 10 });
+     * tree.insert(rect1);
+     * tree.insert(rect2);
+     * rect1.x = 75;
+     * rect1.y = 75;
+     * tree.update(rect1);
+     * ```
+     * @example Bulk fast update and final cleanup:
+     * ```javascript
+     * const tree = new Quadtree({ width: 100, height: 100 });
+     * const rects = [];
+     *  for(let i=0; i<20; i++) {
+     *    rects[i] = new Rectangle({ x: 20, y: 20, width: 20, height: 20 });
+     *    tree.insert(rects[i]);
+     *  }
+     *  for(let i=rects.length-1; i>0; i--) {
+     *    rects[i].x = 20 + Math.random()*60;
+     *    rects[i].y = 20 + Math.random()*60;
+     *    //fast=true – just re-insert the object (may leaves vacant subnodes)
+     *    //fast=false – cleanup empty subnodes (default)
+     *    const fast = (i !== 0);
+     *    tree.update(rects[i], fast);
+     *  }
+     * ```
+     *
+     * @param obj - Object to be updated.
+     * @param fast - Set to true to increase performance temporarily by preventing cleanup of empty subnodes (optional, default: false).
+     */
+    update(obj: ObjectsType, fast?: boolean): void;
     /**
      * The opposite of a split: try to merge and dissolve subnodes.
      * @beta